pyproject.toml
[tool.poetry]
name = "wrime-bert"
version = "0.1.0"
description = ""
authors = ["nakatuba <nakatuba0626@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fugashi = "^1.1.1"
hydra-core = "^1.1.1"
ipadic = "^1.0.0"
matplotlib = "^3.4.3"
omegaconf = "^2.1.1"
pandas = "^1.3.3"
scikit-learn = "^0.24.2"
torch = "^1.9.1"
transformers = "^4.10.2"
wandb = "^0.12.6"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
jupyterlab = "^3.2.0"
jupyterlab-code-formatter = "^1.4.10"
jupyterlab-kite = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
